+++ template.js
@@ -13,7 +13,10 @@ const colors = {
     darkGray: "#404040",
     gray: "#6E6E6E",
     lightGray: "#CECECE",
+    lighterGray: "#D6D6D6",
+    whiteGray: "#DEDEDE",
     white: "#E5E5E5",
+    realwhite: "#FFFFFF",
 
     yellow: "#CCC47A",
     pink: "#E52E71",
@@ -23,24 +26,57 @@ const colors = {
     purple: "#A082D9",
 
     transparent: "#ffffff00",
+};
+
+const dawn = {
+    type: "light",
+
+    // Background colors
+    darkBackground: "white",
+    background: "realwhite",
+    lightBackground: "whiteGray",
+    highlightBackground: "black15",
+    activeBackground: "black25",
+    selectionBackground: "yellow40",
+    inactiveSelectionBackground: "yellow20",
+
+    border: "lightGray",
+
+    // Foregrounds
+    brightForeground: "black",
+    foreground: "darkGray",
+    inactiveForeground: "gray",
+
+    ...colors,
+
+    yellow: "#7e7630",
+    // pink: "",
+    green: "#618122",
+    cyan: "#24798a",
+    // orange: "",
+    purple: "#673abb",
+};
+
+const dusk = {
+    type: "dark",
 
     // Background colors
     darkBackground: "black",
     background: "blackGray",
     lightBackground: "darkerGray",
-    highlightBackground: "darkGray",
+    highlightBackground: "white29",
+    activeBackground: "white40",
+    selectionBackground: "yellow40",
+    inactiveSelectionBackground: "yellow20",
 
     border: "darkGray",
-    matchBorder: "yellow",
 
     // Foregrounds
     brightForeground: "white",
     foreground: "lightGray",
     inactiveForeground: "gray",
 
-    // These colors are used for indicating something,
-    // and will have a bright color.
-    untested: "#FF0000",
+    ...colors,
 };
 
 /**
@@ -52,22 +88,46 @@ const colors = {
  * @param {string} value Value of the color's name.
  */
 function resolveColor(key, value) {
-    if (value in colors) {
-        return resolveColor(value, colors[value]);
+    let valueFormat = /^([a-zA-Z]+)(\d{2})?$/;
+    let matches = value.match(valueFormat);
+    if (matches === null) {
+        // Hex color.
+        return value;
+    }
+
+    let [_fullMatch, colorName, opacity, ..._] = matches;
+    if (!(colorName in colors)) {
+        return value;
+    }
+
+    let resolvedColor = resolveColor(value, colors[colorName]);
+    if(opacity === undefined) {
+        return resolvedColor;
+    }
+    if (colorName in colors) {
+        return resolveColor(value, colors[colorName]) + opacity;
     }
     return value;
 }
+
+const themeTemplates = {
+    "Rubecula": dusk,
+    "Rubecula Dawn": dawn
+};
 
 // Generate a sed replacement command that will replace the colors in the
 // template color scheme and save it as a different file.
 //
 // Example:
 // sed -e "s/\": \"black/\": \"#141414/" themes/Rubecula.template.json > themes/Rubecula.json
-var command = "sed ";
-Object.entries(colors).forEach(([key, value]) => {
-    const color = resolveColor(key, value);
-    command += `-e "s/\\": \\"${key}/\\": \\"${color}/" `;
-});
-command += "themes/Rubecula.template.json > themes/Rubecula.json";
 
-exec(command);
+Object.entries(themeTemplates).forEach(([templateName, templateColors]) => {
+    var command = "sed ";
+    Object.entries(templateColors).forEach(([key, value]) => {
+        const color = resolveColor(key, value);
+        command += `-e "s/\\": \\"${key}/\\": \\"${color}/" `;
+    });
+    command += `"themes/Rubecula.template.json" > "themes/${templateName}.json"`;
+
+    exec(command);
+});